Login/Create Interract
The Create New User page has correct fields and can access user ID etc. The control on the login page takes user to the create new user page. Upon creating an account, it takes the user back to the login page (not necessarily desired effect, but will do for now)