Delay between account creation and login.
Displaying feedback once account is created and the app logging in. This is done in a small thread which will only run once. Also added different strings to the lang properties. Some H/coded strings have found their way in, but these can be fixed later as its not a priority.